+++ src/postfixevaluator/operator/AddOperator.java
@@ -27,7 +27,7 @@ public class AddOperator extends Operator {
      * @return - sum of a + b
      */
     @Override
-    public double evaluate(double a, double b) {
+    public int evaluate(int a, int b) {
         return a + b;
     }
