+++ src/postfixevaluator/operator/SubOperator.java
@@ -28,7 +28,7 @@ public class SubOperator extends Operator {
      * @return - difference of (a - b) 
      */
     @Override
-    public double evaluate(double a, double b) {
+    public int evaluate(int a, int b) {
         return a - b;
     }
